src/main/java/com/polypro/duan1quanlynongsan/Model/PhieuNhapa.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.polypro.duan1quanlynongsan.Model;

import java.util.Date;

/**
 *
 * @author Admin
 */
public class PhieuNhapa {
      private String maDN;
  
    private Date ngayN;
   private String maNV;

    private String maNS;

    private int SoLuong;

    public String getMaDN() {
        return maDN;
    }

    public void setMaDN(String maDN) {
        this.maDN = maDN;
    }

    public Date getNgayN() {
        return ngayN;
    }

    public void setNgayN(Date ngayN) {
        this.ngayN = ngayN;
    }

    public String getMaNV() {
        return maNV;
    }

    public void setMaNV(String maNV) {
        this.maNV = maNV;
    }

    public String getMaNS() {
        return maNS;
    }

    public void setMaNS(String maNS) {
        this.maNS = maNS;
    }

    public int getSoLuong() {
        return SoLuong;
    }

    public void setSoLuong(int SoLuong) {
        this.SoLuong = SoLuong;
    }


}